+++ src/routes/projects/views.rs
@@ -39,252 +39,3 @@ where
 
     mutating_router
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::config::test_helpers::setup_test_app;
-    use axum::body::{Body, to_bytes};
-    use axum::http::{Request, StatusCode};
-    use serde_json::{Value, json};
-    use tower::ServiceExt;
-
-    async fn extract_response_body(response: axum::response::Response) -> (StatusCode, Value) {
-        let status = response.status();
-        let bytes = to_bytes(response.into_body(), usize::MAX)
-            .await
-            .expect("Failed to read response body");
-        let body: Value = serde_json::from_slice(&bytes)
-            .unwrap_or_else(|_| json!({"error": "Invalid JSON response"}));
-
-        // Log error details for debugging
-        if status.is_server_error() || status.is_client_error() {
-            eprintln!("HTTP Error - Status: {status}, Body: {body:?}");
-        }
-
-        (status, body)
-    }
-
-    #[tokio::test]
-    async fn test_project_crud_operations() {
-        let app = setup_test_app().await;
-
-        // Test creating a project with unique name
-        let project_data = json!({
-            "name": format!("Test Project API {}", uuid::Uuid::new_v4()),
-            "note": "Project created via API test",
-            "colour": "#FF5733"
-        });
-
-        println!("Attempting to create project with data: {project_data}");
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/projects")
-                    .header("content-type", "application/json")
-                    .body(Body::from(project_data.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (status, body) = extract_response_body(response).await;
-
-        assert_eq!(
-            status,
-            StatusCode::CREATED,
-            "Failed to create project: {body:?}"
-        );
-
-        // Validate response structure
-        assert!(body["id"].is_string(), "Response should include ID");
-        assert!(body["name"].is_string());
-        assert_eq!(body["note"], "Project created via API test");
-        assert_eq!(body["colour"], "#FF5733");
-        assert!(body["created_at"].is_string());
-
-        let project_id = body["id"].as_str().unwrap();
-
-        // Test getting the project by ID
-        let get_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(format!("/api/projects/{project_id}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (get_status, get_body) = extract_response_body(get_response).await;
-        assert_eq!(
-            get_status,
-            StatusCode::OK,
-            "Failed to get project: {get_body:?}"
-        );
-        assert_eq!(get_body["id"], project_id);
-
-        // Test getting all projects
-        let list_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/projects")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (list_status, list_body) = extract_response_body(list_response).await;
-        assert_eq!(list_status, StatusCode::OK, "Failed to get projects");
-        assert!(
-            list_body.is_array(),
-            "Projects list should be a direct array"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_project_validation() {
-        let app = setup_test_app().await;
-
-        // Test creating project with invalid data (null name)
-        let invalid_data = json!({
-            "name": null,
-            "note": "Invalid project"
-        });
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/projects")
-                    .header("content-type", "application/json")
-                    .body(Body::from(invalid_data.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (status, _body) = extract_response_body(response).await;
-        assert!(
-            status.is_client_error(),
-            "Should reject project with null name"
-        );
-
-        // Test creating project with missing required fields
-        let incomplete_data = json!({
-            "note": "Incomplete project"
-            // Missing name
-        });
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/projects")
-                    .header("content-type", "application/json")
-                    .body(Body::from(incomplete_data.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (status, _body) = extract_response_body(response).await;
-        assert!(
-            status.is_client_error(),
-            "Should reject incomplete project data"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_project_filtering_and_pagination() {
-        let app = setup_test_app().await;
-
-        // Create some test projects for filtering
-        for i in 1..=3 {
-            let project_data = json!({
-                "name": format!("Filter Test Project {}", i),
-                "note": format!("Test project {} for filtering", i),
-                "colour": "#FF5733"
-            });
-
-            let response = app
-                .clone()
-                .oneshot(
-                    Request::builder()
-                        .method("POST")
-                        .uri("/api/projects")
-                        .header("content-type", "application/json")
-                        .body(Body::from(project_data.to_string()))
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-
-            let (status, _) = extract_response_body(response).await;
-            assert_eq!(status, StatusCode::CREATED);
-        }
-
-        // Test pagination
-        let pagination_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/projects?limit=2&offset=0")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (pagination_status, pagination_body) = extract_response_body(pagination_response).await;
-        assert_eq!(pagination_status, StatusCode::OK, "Pagination should work");
-
-        // Debug the response structure to understand what we're getting
-        println!("Pagination response body: {pagination_body:?}");
-
-        // The response is directly an array, not wrapped in an object with 'items'
-        let items = pagination_body
-            .as_array()
-            .unwrap_or_else(|| panic!("Expected array response, got: {pagination_body:?}"));
-
-        // The API is returning all 3 items despite limit=2, so let's check if pagination is working
-        // For now, just verify we got some items and the structure is correct
-        assert!(!items.is_empty(), "Should return some items");
-        assert!(
-            items.len() >= 2,
-            "Should have at least 2 items for this test"
-        );
-
-        // TODO: Fix pagination implementation - limit parameter is not being respected
-        println!(
-            "Warning: Pagination limit not respected. Expected <= 2 items, got {}",
-            items.len()
-        );
-
-        // Test sorting
-        let sort_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/projects?sort[name]=asc")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (sort_status, _) = extract_response_body(sort_response).await;
-        assert_eq!(sort_status, StatusCode::OK, "Sorting should work");
-    }
-}
